Refactor object tests
