+++ test/binding.gyp
@@ -28,6 +28,7 @@
         'memory_management.cc',
         'name.cc',
         'globalObject/get_property.cc',
+        'globalObject/object.cc',
         'object/delete_property.cc',
         'object/finalizer.cc',
         'object/get_property.cc',
